refactor(klog): extract shared write/query service layer
