+++ src/kernel/klog/src/network/server.rs
@@ -1,11 +1,10 @@
-use super::KDataClient;
 use super::request::{
-    KLOG_FORWARD_HOPS_HEADER, KLOG_FORWARDED_BY_HEADER, KLogAdminRequestType, KLogAppendRequest,
-    KLogAppendResponse, KLogClusterStateResponse, KLogDataRequestType, RaftRequest,
-    RaftRequestType, RaftResponse,
+    KLogAdminRequestType, KLogAppendRequest, KLogClusterStateResponse, KLogDataRequestType,
+    RaftRequest, RaftRequestType, RaftResponse,
 };
+use crate::service::KLogWriteService;
 use crate::state_store::KLogStateStoreManagerRef;
-use crate::{KLogEntry, KLogRequest, KLogResponse, KNode, KNodeId, KRaftRef};
+use crate::{KNode, KNodeId, KRaftRef};
 use axum::Json;
 use axum::Router;
 use axum::body::Bytes;
@@ -21,7 +20,6 @@ use std::collections::BTreeSet;
 use std::future::Future;
 use std::net::SocketAddr;
 use std::time::Duration;
-use std::time::{SystemTime, UNIX_EPOCH};
 use tower::BoxError;
 use tower::ServiceBuilder;
 use tower::limit::ConcurrencyLimitLayer;
@@ -36,9 +34,6 @@ const CONTROL_RPC_CONCURRENCY_LIMIT: usize = 128;
 const SNAPSHOT_RPC_CONCURRENCY_LIMIT: usize = 8;
 const CONTROL_RPC_TIMEOUT_MS: u64 = 3_000;
 const SNAPSHOT_RPC_TIMEOUT_MS: u64 = 30_000;
-const DATA_APPEND_MAX_MESSAGE_BYTES: usize = 64 * 1024;
-const DATA_APPEND_MAX_FORWARD_HOPS: u32 = 2;
-
 #[derive(Debug, Deserialize)]
 struct AddLearnerQuery {
     node_id: KNodeId,
@@ -62,8 +57,7 @@ struct RemoveLearnerQuery {
 #[derive(Clone)]
 struct KNetworkServerState {
     raft: KRaftRef,
-    state_store_manager: Option<KLogStateStoreManagerRef>,
-    data_client: KDataClient,
+    write_service: Option<KLogWriteService>,
     admin_local_only: bool,
     cluster_name: String,
     cluster_id: String,
@@ -119,8 +113,9 @@ impl KNetworkServer {
     {
         let state = KNetworkServerState {
             raft: self.raft.clone(),
-            state_store_manager: self.state_store_manager.clone(),
-            data_client: KDataClient::new(),
+            write_service: self.state_store_manager.clone().map(|state_store_manager| {
+                KLogWriteService::new("KNetworkServer", self.raft.clone(), state_store_manager)
+            }),
             admin_local_only: self.admin_local_only,
             cluster_name: self.cluster_name.clone(),
             cluster_id: self.cluster_id.clone(),
@@ -375,187 +370,16 @@ impl KNetworkServer {
         headers: HeaderMap,
         Json(req): Json<KLogAppendRequest>,
     ) -> Response {
-        match Self::process_data_append(&state, &headers, req).await {
-            Ok(resp) => (StatusCode::OK, Json(resp)).into_response(),
-            Err((status, msg)) => Self::error_response(status, msg),
-        }
-    }
-
-    async fn process_data_append(
-        state: &KNetworkServerState,
-        headers: &HeaderMap,
-        req: KLogAppendRequest,
-    ) -> Result<KLogAppendResponse, (StatusCode, String)> {
-        if req.message.trim().is_empty() {
-            let msg = "KNetworkServer data append rejected: empty message".to_string();
-            error!("{}", msg);
-            return Err((StatusCode::BAD_REQUEST, msg));
-        }
-
-        if req.message.len() > DATA_APPEND_MAX_MESSAGE_BYTES {
-            let msg = format!(
-                "KNetworkServer data append rejected: message too large, bytes={}, max_bytes={}",
-                req.message.len(),
-                DATA_APPEND_MAX_MESSAGE_BYTES
-            );
-            error!("{}", msg);
-            return Err((StatusCode::PAYLOAD_TOO_LARGE, msg));
-        }
-
-        let forward_hops = Self::parse_forward_hops(headers).map_err(|msg| {
-            error!("{}", msg);
-            (StatusCode::BAD_REQUEST, msg)
-        })?;
-        let forwarded_by = headers
-            .get(KLOG_FORWARDED_BY_HEADER)
-            .and_then(|v| v.to_str().ok())
-            .unwrap_or("-");
-
-        if forward_hops > DATA_APPEND_MAX_FORWARD_HOPS {
-            let msg = format!(
-                "KNetworkServer data append rejected: too many forward hops, hops={}, max_hops={}, forwarded_by={}",
-                forward_hops, DATA_APPEND_MAX_FORWARD_HOPS, forwarded_by
-            );
-            error!("{}", msg);
-            return Err((StatusCode::BAD_GATEWAY, msg));
-        }
-
-        let Some(state_store_manager) = state.state_store_manager.as_ref() else {
+        let Some(write_service) = state.write_service.as_ref() else {
             let msg = "KNetworkServer data append rejected: state store manager is not configured"
                 .to_string();
             error!("{}", msg);
-            return Err((StatusCode::INTERNAL_SERVER_ERROR, msg));
+            return Self::error_response(StatusCode::INTERNAL_SERVER_ERROR, msg);
         };
 
-        let metrics = state.raft.metrics().borrow().clone();
-        let local_node_id = metrics.id;
-        let req = KLogAppendRequest {
-            message: req.message,
-            timestamp: req.timestamp.or_else(|| Some(Self::now_millis())),
-            node_id: req.node_id.or(Some(local_node_id)),
-        };
-
-        let item = state_store_manager.prepare_append_entry(KLogEntry {
-            id: 0,
-            timestamp: req.timestamp.unwrap_or(0),
-            node_id: req.node_id.unwrap_or(local_node_id),
-            message: req.message.clone(),
-        });
-        let requested_id = item.id;
-
-        info!(
-            "KNetworkServer data append request: id={}, timestamp={}, node_id={}, msg_len={}, local_node_id={}, current_leader={:?}, forward_hops={}, forwarded_by={}",
-            item.id,
-            item.timestamp,
-            item.node_id,
-            item.message.len(),
-            local_node_id,
-            metrics.current_leader,
-            forward_hops,
-            forwarded_by
-        );
-
-        match state
-            .raft
-            .client_write(KLogRequest::AppendLog { item })
-            .await
-        {
-            Ok(resp) => match resp.data {
-                KLogResponse::AppendOk { id } => {
-                    info!("KNetworkServer data append committed: id={}", id);
-                    Ok(KLogAppendResponse { id })
-                }
-                KLogResponse::Err(err_msg) => {
-                    let msg = format!(
-                        "KNetworkServer data append failed in state machine: requested_id={}, err={}",
-                        requested_id, err_msg
-                    );
-                    error!("{}", msg);
-                    Err((StatusCode::INTERNAL_SERVER_ERROR, msg))
-                }
-                other => {
-                    let msg = format!(
-                        "KNetworkServer data append unexpected response: requested_id={}, response={:?}",
-                        requested_id, other
-                    );
-                    error!("{}", msg);
-                    Err((StatusCode::INTERNAL_SERVER_ERROR, msg))
-                }
-            },
-            Err(err) => {
-                if let Some(forward) = err.forward_to_leader::<KNode>() {
-                    if forward_hops >= DATA_APPEND_MAX_FORWARD_HOPS {
-                        let msg = format!(
-                            "KNetworkServer data append forward aborted due to hop limit: local_node_id={}, requested_id={}, leader_id={:?}, leader_node={:?}, hops={}, max_hops={}",
-                            local_node_id,
-                            requested_id,
-                            forward.leader_id,
-                            forward.leader_node,
-                            forward_hops,
-                            DATA_APPEND_MAX_FORWARD_HOPS
-                        );
-                        error!("{}", msg);
-                        return Err((StatusCode::BAD_GATEWAY, msg));
-                    }
-
-                    let leader_node = forward.leader_node.clone().or_else(|| {
-                        forward.leader_id.and_then(|leader_id| {
-                            metrics
-                                .membership_config
-                                .nodes()
-                                .find_map(|(id, node)| (*id == leader_id).then_some(node.clone()))
-                        })
-                    });
-                    let Some(leader_node) = leader_node else {
-                        let msg = format!(
-                            "KNetworkServer data append can not resolve leader node for forwarding: local_node_id={}, requested_id={}, leader_id={:?}",
-                            local_node_id, requested_id, forward.leader_id
-                        );
-                        warn!("{}", msg);
-                        return Err((StatusCode::SERVICE_UNAVAILABLE, msg));
-                    };
-
-                    let target_hops = forward_hops + 1;
-                    warn!(
-                        "KNetworkServer data append forwarding to leader: local_node_id={}, requested_id={}, leader_id={}, leader_addr={}:{}, hops={} -> {}",
-                        local_node_id,
-                        requested_id,
-                        leader_node.id,
-                        leader_node.addr,
-                        leader_node.port,
-                        forward_hops,
-                        target_hops
-                    );
-                    match state
-                        .data_client
-                        .append_to_node(&leader_node, &req, target_hops, local_node_id)
-                        .await
-                    {
-                        Ok(resp) => {
-                            info!(
-                                "KNetworkServer data append forwarded and committed: local_node_id={}, requested_id={}, committed_id={}, leader_id={}, hops={}",
-                                local_node_id, requested_id, resp.id, leader_node.id, target_hops
-                            );
-                            Ok(resp)
-                        }
-                        Err(forward_err) => {
-                            let msg = format!(
-                                "KNetworkServer data append forward failed: local_node_id={}, requested_id={}, leader_id={}, err={}",
-                                local_node_id, requested_id, leader_node.id, forward_err
-                            );
-                            error!("{}", msg);
-                            Err((StatusCode::BAD_GATEWAY, msg))
-                        }
-                    }
-                } else {
-                    let msg = format!(
-                        "KNetworkServer data append raft client_write failed: requested_id={}, err={}",
-                        requested_id, err
-                    );
-                    error!("{}", msg);
-                    Err((StatusCode::INTERNAL_SERVER_ERROR, msg))
-                }
-            }
+        match write_service.append(&headers, req).await {
+            Ok(resp) => (StatusCode::OK, Json(resp)).into_response(),
+            Err((status, msg)) => Self::error_response(status, msg),
         }
     }
 
@@ -819,31 +643,6 @@ impl KNetworkServer {
         Some(Self::error_response(StatusCode::FORBIDDEN, msg))
     }
 
-    fn parse_forward_hops(headers: &HeaderMap) -> Result<u32, String> {
-        let Some(raw) = headers.get(KLOG_FORWARD_HOPS_HEADER) else {
-            return Ok(0);
-        };
-        let raw = raw.to_str().map_err(|e| {
-            format!(
-                "KNetworkServer data append invalid {} header utf8: {}",
-                KLOG_FORWARD_HOPS_HEADER, e
-            )
-        })?;
-        raw.parse::<u32>().map_err(|e| {
-            format!(
-                "KNetworkServer data append invalid {} header '{}': {}",
-                KLOG_FORWARD_HOPS_HEADER, raw, e
-            )
-        })
-    }
-
-    fn now_millis() -> u64 {
-        SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .map(|d| d.as_millis() as u64)
-            .unwrap_or(0)
-    }
-
     fn error_response(status: StatusCode, msg: String) -> Response {
         (status, msg).into_response()
     }
@@ -871,9 +670,7 @@ fn parse_voter_ids_csv(raw: &str) -> Result<Vec<KNodeId>, String> {
 
 #[cfg(test)]
 mod tests {
-    use super::{KNetworkServer, parse_voter_ids_csv};
-    use crate::network::KLOG_FORWARD_HOPS_HEADER;
-    use axum::http::HeaderMap;
+    use super::parse_voter_ids_csv;
 
     #[test]
     fn test_parse_voter_ids_csv_ok() {
@@ -891,32 +688,5 @@ mod tests {
     fn test_parse_voter_ids_csv_invalid_rejected() {
         let err = parse_voter_ids_csv("1,a").expect_err("invalid voter should fail");
         assert!(err.contains("invalid node id"));
-    }
-
-    #[test]
-    fn test_parse_forward_hops_default_zero() {
-        let headers = HeaderMap::new();
-        let hops = KNetworkServer::parse_forward_hops(&headers).expect("parse hops");
-        assert_eq!(hops, 0);
-    }
-
-    #[test]
-    fn test_parse_forward_hops_ok() {
-        let mut headers = HeaderMap::new();
-        headers.insert(KLOG_FORWARD_HOPS_HEADER, "2".parse().expect("header value"));
-        let hops = KNetworkServer::parse_forward_hops(&headers).expect("parse hops");
-        assert_eq!(hops, 2);
-    }
-
-    #[test]
-    fn test_parse_forward_hops_invalid_rejected() {
-        let mut headers = HeaderMap::new();
-        headers.insert(
-            KLOG_FORWARD_HOPS_HEADER,
-            "abc".parse().expect("header value"),
-        );
-        let err = KNetworkServer::parse_forward_hops(&headers).expect_err("invalid hops");
-        assert!(err.contains("invalid"));
-        assert!(err.contains(KLOG_FORWARD_HOPS_HEADER));
     }
 }
